requirements.txt
requests==2.32.3
streamlit==1.38.0
pandas==2.2.2
python-dotenv==1.0.1
spotipy==2.25.1